part 202 completed

+++ database/migrations/2022_05_23_190839_add_polymorph_to_images_table.php
@@ -28,7 +28,7 @@ class AddPolymorphToImagesTable extends Migration
     public function down()
     {
         Schema::table('images', function (Blueprint $table) {
-            $table->unsignedInteger('blog_post_id')->nullable();
+            $table->unsignedBigInteger('blog_post_id')->nullable();
             $table->dropMorphs('imageable');
         });
     }
